==> scr/00_install_libraries.R <==
renv::install(c("bioc::fgsea","bioc::AnnotationDbi","bioc::apeglm","bioc::tximport","bioc::tximportData","bioc::vsn","bioc::limma","bioc::org.Hs.eg.db","bioc::biomaRt","bioc::IRanges","bioc::GenomicFeatures","bioc::GSEABase","bioc::DESeq2","bioc::IHW","bioc::ashr","PoiClaClu"))

renv::install(c("tidyverse","hexbin","patchwork","ggrepel","enrichR","msigdbr","viridis"))

renv::install(c("jokergoo/ComplexHeatmap"))

renv::install("pheatmap")

renv::install("bioc::AnnotationHub")
renv::install("GGally")
renv::install("RNAseqQC")
renv::install("edgeR")

renv::install(c("UpSetR","gplots","pals"))
